apps/backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.6.7
bcrypt==4.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.1
# uvloop removed for Windows compatibility (uvloop does not support Windows)
watchfiles==1.1.1
websockets==16.0
itsdangerous==2.1.2
